restructure sqs
