+++ src/main/java/org/sustain/handlers/spark/regression/RandomForestRegressionHandler.java
@@ -1,4 +1,4 @@
-package org.sustain.handlers.regression;
+package org.sustain.handlers.spark.regression;
 
 import io.grpc.stub.StreamObserver;
 import org.apache.logging.log4j.LogManager;
@@ -6,16 +6,16 @@ import org.apache.logging.log4j.Logger;
 import org.sustain.ModelRequest;
 import org.sustain.ModelResponse;
 import org.sustain.SparkManager;
-import org.sustain.handlers.tasks.GradientBoostRegressionTask;
-import org.sustain.handlers.tasks.RegressionTask;
+import org.sustain.tasks.spark.regression.RandomForestRegressionTask;
+import org.sustain.tasks.spark.regression.RegressionTask;
 
 import java.util.List;
 
-public class GradientBoostRegressionQueryHandler extends RegressionQueryHandler {
+public class RandomForestRegressionHandler extends RegressionHandler {
 
-    private static final Logger log = LogManager.getLogger(GradientBoostRegressionQueryHandler.class);
+    private static final Logger log = LogManager.getLogger(RandomForestRegressionHandler.class);
 
-    public GradientBoostRegressionQueryHandler(ModelRequest request,
+    public RandomForestRegressionHandler(ModelRequest request,
                                         StreamObserver<ModelResponse> responseObserver,
                                         SparkManager sparkManager) {
         super(request, responseObserver, sparkManager);
@@ -23,11 +23,11 @@ public class GradientBoostRegressionQueryHandler extends RegressionQueryHandler 
 
     @Override
     public boolean hasAppropriateRegressionRequest(ModelRequest modelRequest) {
-        return modelRequest.hasGBoostRegressionRequest();
+        return modelRequest.hasRForestRegressionRequest();
     }
 
     @Override
     public RegressionTask createRegressionTask(List<String> gisJoinBatch) {
-        return new GradientBoostRegressionTask(this.request, gisJoinBatch);
+        return new RandomForestRegressionTask(this.request, gisJoinBatch);
     }
 }
