+++ src/main/java/org/sustain/modeling/spark/regression/SustainRegressionModel.java
@@ -1,4 +1,4 @@
-package org.sustain.modeling;
+package org.sustain.modeling.spark.regression;
 
 
 import org.apache.spark.ml.regression.RegressionModel;
